tests: Remove self_impl.h
